+++ src/AutoDeconstruct.Tests/AnalyzeAttributeUsageAnalyzerTests.cs
@@ -2,11 +2,60 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Testing;
 using NUnit.Framework;
+using System.Globalization;
 
 namespace AutoDeconstruct.Tests;
 
 internal static class AnalyzeAttributeUsageAnalyzerTests
 {
+	[Test]
+	public static void VerifySupportedDiagnostics()
+	{
+		var analyzer = new AnalyzeAttributeUsageAnalyzer();
+		var diagnostics = analyzer.SupportedDiagnostics;
+
+		using (Assert.EnterMultipleScope())
+		{
+			Assert.That(diagnostics, Has.Length.EqualTo(3), nameof(diagnostics.Length));
+			Assert.That(diagnostics.All(_ => _.Category == DescriptorConstants.Usage), Is.True, nameof(diagnostics));
+			Assert.That(diagnostics.All(_ => _.DefaultSeverity == DiagnosticSeverity.Error), Is.True, nameof(diagnostics));
+			Assert.That(diagnostics.All(_ => _.IsEnabledByDefault), Is.True, nameof(diagnostics));
+
+			var noAccessiblePropertiesDiagnostic = diagnostics.Single(_ => _.Id == NoAccessiblePropertiesDescriptor.Id);
+			Assert.That(noAccessiblePropertiesDiagnostic.Title.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(NoAccessiblePropertiesDescriptor.Title),
+				nameof(DiagnosticDescriptor.Title));
+			Assert.That(noAccessiblePropertiesDiagnostic.MessageFormat.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(NoAccessiblePropertiesDescriptor.Message),
+				nameof(DiagnosticDescriptor.MessageFormat));
+			Assert.That(noAccessiblePropertiesDiagnostic.HelpLinkUri,
+				Is.EqualTo(HelpUrlBuilder.Build(NoAccessiblePropertiesDescriptor.Id)),
+				nameof(DiagnosticDescriptor.HelpLinkUri));
+
+			var instanceDeconstructDiagnostic = diagnostics.Single(_ => _.Id == InstanceDeconstructExistsDescriptor.Id);
+			Assert.That(instanceDeconstructDiagnostic.Title.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(InstanceDeconstructExistsDescriptor.Title),
+				nameof(DiagnosticDescriptor.Title));
+			Assert.That(instanceDeconstructDiagnostic.MessageFormat.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(InstanceDeconstructExistsDescriptor.Message),
+				nameof(DiagnosticDescriptor.MessageFormat));
+			Assert.That(instanceDeconstructDiagnostic.HelpLinkUri,
+				Is.EqualTo(HelpUrlBuilder.Build(InstanceDeconstructExistsDescriptor.Id)),
+				nameof(DiagnosticDescriptor.HelpLinkUri));
+
+			var extensionDeconstructDiagnostic = diagnostics.Single(_ => _.Id == ExtensionDeconstructExistsDescriptor.Id);
+			Assert.That(extensionDeconstructDiagnostic.Title.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(ExtensionDeconstructExistsDescriptor.Title),
+				nameof(DiagnosticDescriptor.Title));
+			Assert.That(extensionDeconstructDiagnostic.MessageFormat.ToString(CultureInfo.CurrentCulture),
+				Is.EqualTo(ExtensionDeconstructExistsDescriptor.Message),
+				nameof(DiagnosticDescriptor.MessageFormat));
+			Assert.That(extensionDeconstructDiagnostic.HelpLinkUri,
+				Is.EqualTo(HelpUrlBuilder.Build(ExtensionDeconstructExistsDescriptor.Id)),
+				nameof(DiagnosticDescriptor.HelpLinkUri));
+		}
+	}
+
 	[Test]
 	public static async Task AnalyzeWhenAssemblyAttributeIsGoodAsync()
 	{
@@ -52,6 +101,65 @@ internal static class AnalyzeAttributeUsageAnalyzerTests
 	}
 
 	[Test]
+	public static async Task AnalyzeWhenAssemblyAttributeFindsInstanceDeconstructAsync()
+	{
+		var code =
+			"""
+			using AutoDeconstruct;
+			using System;
+
+			[assembly: TargetAutoDeconstruct(typeof(TestSpace.Test))]
+
+			namespace TestSpace
+			{
+				public class Test
+				{ 
+					public string? Name { get; set; }
+					public Guid Id { get; set; }
+
+					public void Deconstruct(out Guid id, out string name) =>
+						(id, name) = (this.Id, this.Name);
+				}
+			}
+			""";
+
+		var diagnostic = new DiagnosticResult(InstanceDeconstructExistsDescriptor.Id, DiagnosticSeverity.Error)
+			.WithSpan(4, 12, 4, 57);
+		await TestAssistants.RunAnalyzerAsync<AnalyzeAttributeUsageAnalyzer>(code, [diagnostic]);
+	}
+
+	[Test]
+	public static async Task AnalyzeWhenAssemblyAttributeFindsExtensionDeconstructAsync()
+	{
+		var code =
+			"""
+			using AutoDeconstruct;
+			using System;
+
+			[assembly: TargetAutoDeconstruct(typeof(TestSpace.Test), SearchForExtensionMethods.Yes)]
+
+			namespace TestSpace
+			{
+				public class Test
+				{ 
+					public string? Name { get; set; }
+					public Guid Id { get; set; }
+				}
+
+				public static class TestExtensions
+				{
+					public static void Deconstruct(this Test self, out Guid id, out string name) =>
+						(id, name) = (self.Id, self.Name);
+				}
+			}
+			""";
+
+		var diagnostic = new DiagnosticResult(ExtensionDeconstructExistsDescriptor.Id, DiagnosticSeverity.Error)
+			.WithSpan(4, 12, 4, 88);
+		await TestAssistants.RunAnalyzerAsync<AnalyzeAttributeUsageAnalyzer>(code, [diagnostic]);
+	}
+
+	[Test]
 	public static async Task AnalyzeWhenTypeAttributeIsGoodAsync()
 	{
 		var code =
@@ -90,6 +198,63 @@ internal static class AnalyzeAttributeUsageAnalyzerTests
 
 		var diagnostic = new DiagnosticResult(NoAccessiblePropertiesDescriptor.Id, DiagnosticSeverity.Error)
 			.WithSpan(6, 3, 6, 18);
+		await TestAssistants.RunAnalyzerAsync<AnalyzeAttributeUsageAnalyzer>(code, [diagnostic]);
+	}
+
+	[Test]
+	public static async Task AnalyzeWhenTypeAttributeFindsInstanceDeconstructAsync()
+	{
+		var code =
+			"""
+			using AutoDeconstruct;
+			using System;
+
+			namespace TestSpace
+			{
+				[AutoDeconstruct]
+				public class Test 
+				{ 
+					public string? Name { get; set; }
+					public Guid Id { get; set; }
+			
+					public void Deconstruct(out Guid id, out string name) =>
+						(id, name) = (this.Id, this.Name);				
+				}
+			}
+			""";
+
+		var diagnostic = new DiagnosticResult(InstanceDeconstructExistsDescriptor.Id, DiagnosticSeverity.Error)
+			.WithSpan(6, 3, 6, 18);
+		await TestAssistants.RunAnalyzerAsync<AnalyzeAttributeUsageAnalyzer>(code, [diagnostic]);
+	}
+
+	[Test]
+	public static async Task AnalyzeWhenTypeAttributeFindsExtensionDeconstructAsync()
+	{
+		var code =
+			"""
+			using AutoDeconstruct;
+			using System;
+
+			namespace TestSpace
+			{
+				[AutoDeconstruct(SearchForExtensionMethods.Yes)]
+				public class Test 
+				{ 
+					public string? Name { get; set; }
+					public Guid Id { get; set; }			
+				}
+
+				public static class TestExtensions
+				{
+					public static void Deconstruct(this Test self, out Guid id, out string name) =>
+						(id, name) = (self.Id, self.Name);				
+				}
+			}
+			""";
+
+		var diagnostic = new DiagnosticResult(ExtensionDeconstructExistsDescriptor.Id, DiagnosticSeverity.Error)
+			.WithSpan(6, 3, 6, 49);
 		await TestAssistants.RunAnalyzerAsync<AnalyzeAttributeUsageAnalyzer>(code, [diagnostic]);
 	}
 }